Implements non-NULL tuple updates
All assignment 1 tests passing, just with a lot of BufferManager
warnings. The general idea is that we find where the updated value
needs to go (which requires a fair bit of work for NULLs), unset
the NULL flag, add/remove space as required, then write the data.
This should complete part 1 of the set.