Implements optimal join planning
Dynamic programming to find optimal join order now works. There's
some messy stuff with NATURAL JOINS that need to be figured out,
but most tests are passing. Ran the query in the set, and it looks
good. Results are consistent with Eric's set from last year and
the solution set (according to Eric), but Donnie wants us to do
some more with NATURAL JOINS.