Implement step 5 (rollback)
Starting from the most recent log record, undo it and write a redo
only record using the redo data. Then, move to the previous log record.
Stop when the beginning of the transaction is reached.

Fixed a bug in forceWAL.

Did some simple testing of rollbacks to see if they work.