Implemented adding and removing rows from indexes.
When adding a new tuple to a table, it is first checked to see if it fails
any constraints. If it does, then the tuple is removed. Otherwise, iterate
through all indexes on the table, and add the corresponding values from
the tuple.

When removing a tuple from a table, iterate through the indexes and remove
the corresponding values.